Further u-boot.deb build support
Building was not working correctly on a clean build.
This change is across two repositories:

    nexus-build
    ti-bdebstrap

On the dev board the result is simply the directory
/usr/lib/u-boot/platform/ containing the u-boot components.

+++ build_iGOS_TMDS64EVM_fs.sh
@@ -92,7 +92,8 @@ if [ ! -f "$BLT" ]; then
     echo "=== I: $0: $TSK BEGIN"
 
     # Symlink everything to the vyos-build/packages directory
-    for a in $(find $ROOTDIR/vyos-build/scripts -type f -name "*.deb")
+    for a in $(find $ROOTDIR/vyos-build/scripts -type f -name "*.deb")\
+             $ROOTDIR/ti-bdebstrap/*.deb
     do
         case "$a" in
         *libsnmp-dev_*64.deb)  # Needed for frr (despite -dev_ pattern)
@@ -112,6 +113,7 @@ if [ ! -f "$BLT" ]; then
             ;;
         esac
 
+        test -f "$a" || continue
         echo "Symlinking package: $a"
         ln -vrfs $a $ROOTDIR/vyos-build/packages/
     done
